feat: rename lexer rule files and update package structure to use lexer.token
