+++ modules/lexer/src/main/kotlin/lexer/token/rules/StringLiteralRule.kt
@@ -1,5 +1,6 @@
-package lexer.rules
+package lexer.token.rules
 
+import lexer.token.TokenRule
 import token.StringLiteralToken
 
 class StringLiteralRule : TokenRule {
